Moving current management to new (old) foolder
Ideally, these will all be cleaned up until only a few files are left.
