+++ old-management/dynamic.py
